+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Security in case of vulnerabilities.
 - Ability to edit project information for conversions and transfers
 - ProjectGroup is created when a project is edited with a new GRN
 - Added `Add, Edit and Delete External Contacts for both conversion and transfer projects.
+- Added `Confirm this transfer has authority to proceed` task
 
 ### Changed
 - Group reference number links to the group on "About the project"
@@ -33,6 +34,7 @@ Security in case of vulnerabilities.
 - GroupReferenceNumberAttribute failed when there was no existing group
 - update "Task - Supplemental funding agreement - transfer - incorrect options"
 - Fixed `External stakeholder kick off` task wording for both conversion and transfer projects. 
+- Fixed Error Summary partial view expect model, pass null wherever not required
 
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-12.746...main) for everything awaiting release